requirements.txt
# MCP Server dependencies
mcp>=1.0.0

# Flinks API client
flinks>=0.1.0a2
requests>=2.31.0

# Data processing and analysis
pandas>=2.0.0
numpy>=1.24.0

# Date/time handling
python-dateutil>=2.8.2

# Configuration and environment
python-dotenv>=1.0.0
pydantic>=2.0.0

# Async support
asyncio

# Development dependencies (uncomment for development)
# pytest>=7.0.0
# pytest-asyncio>=0.21.0
# black>=23.0.0
# flake8>=6.0.0 